Make `import { fabric } from "fabric";` work (2) (#23787)
