+++ types/fabric/test/import.ts
@@ -1,0 +1,2 @@
+import { fabric } from "fabric";
+new fabric.Canvas("C");
